디자인 네모네모하게 수정, 클라우드 코끼리 연결

+++ app/src/main/java/com/edu/lx/onedayworkfinal/offer/recycler_view/HumanResourceRecyclerViewAdapter.java
@@ -65,7 +65,7 @@ class HumanResourceRecyclerViewAdapter extends BaseRecyclerViewAdapter<ManageHum
             double total = manageHumanResourceModel.getTotal();
             if (total != 0) {
                 int reliability = (int) ((workOffCount/total) * 100);
-                String reliabilityStr = reliability + "% ("+workOffCount + "/" + total + ")";
+                String reliabilityStr = reliability + "% ("+(int)workOffCount + "/" + (int)total + ")";
                 if (reliability < 80) {
                     reliabilityView.setTextColor(context.getResources().getColor(R.color.danger,context.getTheme()));
                 } else if (reliability < 90){
